refactoring(Package Structure): updated Package structure
