+++ phase2/src/main/java/GUI/SignUpPage.java
@@ -1,6 +1,6 @@
 package main.java.GUI;
-import main.java.Web.OrganizerSystemManager;
-import main.java.Web.ParticipantSystemManager;
+import main.java.SystemManagers.OrganizerSystemManager;
+import main.java.SystemManagers.ParticipantSystemManager;
 
 import javax.swing.*;
 import java.awt.*;
