+++ phase2/src/main/java/RaffleComponent/RaffleRuleSetterUseCase.java
@@ -1,15 +1,10 @@
 package main.java.RaffleComponent;
 
 import main.java.DatabaseRe.AccessData;
+import main.java.DatabaseRe.DataAccessPoint;
+import main.java.DatabaseRe.DataProviderPoint;
 import main.java.DatabaseRe.ProvideData;
-import main.java.Helpers.PackageRaffleEntityInstance;
-import main.java.RaffleComponent.OrganizerRaffleEntity;
-import main.java.database.AddOrganizer;
-import main.java.database.DataExtractor;
 
-import java.awt.image.AreaAveragingScaleFilter;
-import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.sql.SQLException;
 import java.text.ParseException;
 import java.time.LocalDate;
