+++ phase2/src/main/java/Web/UserController.java
@@ -1,4 +1,4 @@
-package main.java.UserWeb;
+package main.java.Web;
 
 import main.java.database.*;
 import main.java.UserComponent.GetOrganizerUseCase;
